+++ apps/indexer/tests/sovereign-nature-identifier.test.ts
@@ -29,19 +29,11 @@ import { SUBGRAPH_ENTITY_NAME } from '@sni/constants';
 //TODO: Refactor constants to object
 //TODO: Add URI postfix to mocked URIs
 import {
-  DERIVATIVE_METADATA_SCHEMA,
-  DERIVATIVE_METADATA_SCHEMA_DIGEST,
-  INITIAL_COMPUTE_URI,
-  INITIAL_DATA_URI,
-  INITIAL_STATUS,
   INITIAL_TOKEN_ID,
-  INITIAL_TOKEN_URI,
   MINTER_ADDRESS,
   OWNER_ADDRESS,
   SECOND_OWNER_ADDRESS,
   TEMP_TOKEN_ID,
-  TOKEN_URI_SCHEMA,
-  TOKEN_URI_SCHEMA_DIGEST,
   UPDATED_COMPUTE_URI,
   UPDATED_DATA_URI,
   UPDATED_STATUS,
@@ -96,116 +88,6 @@ describe('SNI Indexer', () => {
       INITIAL_TOKEN_ID_INT.toHex(),
       'updatedAt',
       transferEvent.block.timestamp.toString()
-    );
-  });
-
-  test('Handles Transfer event', () => {
-    const transferEvent = createTransferEvent(
-      MINTER,
-      OWNER,
-      INITIAL_TOKEN_ID_INT
-    );
-    handleTransfer(transferEvent);
-
-    const tokenId = INITIAL_TOKEN_ID_INT.toHex();
-
-    // Base fields
-    assert.fieldEquals(SUBGRAPH_ENTITY_NAME, tokenId, 'owner', OWNER.toHex());
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'tokenId',
-      INITIAL_TOKEN_ID_INT.toString()
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'tokenURI',
-      INITIAL_TOKEN_URI
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'dataURI',
-      INITIAL_DATA_URI
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'computeURI',
-      INITIAL_COMPUTE_URI
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'status',
-      INITIAL_STATUS.toString()
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'tokenMetadataSchemaURI',
-      TOKEN_URI_SCHEMA
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'tokenMetadataSchemaDigest',
-      TOKEN_URI_SCHEMA_DIGEST
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'derivativeMetadataSchemaURI',
-      DERIVATIVE_METADATA_SCHEMA
-    );
-
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'derivativeMetadataSchemaDigest',
-      DERIVATIVE_METADATA_SCHEMA_DIGEST
-    );
-
-    // JSON metadata fields from initial.json
-    assert.fieldEquals(SUBGRAPH_ENTITY_NAME, tokenId, 'name', 'Scarface');
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'description',
-      'Test lion identifier'
-    );
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'image',
-      'ipfs://bafybeihs6qouvmo4pnjozlrdmgic3b4nav6rrswc3tobgclrrvtwsa47oe/blob'
-    );
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'statusDescription',
-      '{ "0": "Alive", "1": "Dead" }'
-    );
-    assert.fieldEquals(SUBGRAPH_ENTITY_NAME, tokenId, 'taxonId', 'itis:183803');
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'conservationStatus',
-      'VU'
-    );
-    assert.fieldEquals(
-      SUBGRAPH_ENTITY_NAME,
-      tokenId,
-      'geometry',
-      'POINT(5.9559 47.8084)'
     );
   });
